Adjust commands related to config access/reload
